chore: move examples to different file
